requirements.txt
pydantic==2.10.6
pyyaml>=6.0
jinja2>=3.1.0
python-dotenv>=1.0.0
openai>=1.30.0
streamlit>=1.28.0
playwright>=1.40.0